move extract_testing_args out of __init__ as it doens't import properly
