+++ test/cim/iec61968/assetinfo/test_transformer_tank_info.py
@@ -7,7 +7,7 @@
 from hypothesis import given
 from hypothesis.strategies import lists, builds
 
-from test.cim import extract_testing_args
+from test.cim.extract_testing_args import extract_testing_args
 from test.cim.collection_validator import validate_collection_unordered
 from test.cim.iec61968.assets.test_asset_info import asset_info_kwargs, verify_asset_info_constructor_default, \
     verify_asset_info_constructor_kwargs, verify_asset_info_constructor_args, asset_info_args
