+++ test/cim/iec61968/assets/test_pole.py
@@ -6,7 +6,7 @@
 from hypothesis import given
 from hypothesis.strategies import text, lists, builds
 
-from test.cim import extract_testing_args
+from test.cim.extract_testing_args import extract_testing_args
 from test.cim.collection_validator import validate_collection_unordered
 from test.cim.iec61968.assets.test_structure import structure_kwargs, verify_structure_constructor_default, \
     verify_structure_constructor_kwargs, verify_structure_constructor_args, structure_args
