+++ test/cim/iec61968/common/test_town_detail.py
@@ -8,6 +8,7 @@ from hypothesis import given
 from hypothesis.strategies import text
 from test.cim import extract_testing_args
 from test.cim.cim_creators import ALPHANUM, TEXT_MAX_SIZE
+from test.cim.extract_testing_args import extract_testing_args
 from zepben.evolve import TownDetail
 from zepben.evolve.model.cim.iec61968.common.create_common_components import create_town_detail
