+++ test/cim/iec61968/metering/test_usage_point.py
@@ -7,6 +7,7 @@
 from hypothesis import given
 from hypothesis.strategies import builds, lists, booleans, text
 from test.cim import extract_testing_args
+from test.cim.extract_testing_args import extract_testing_args
 from test.cim.cim_creators import ALPHANUM, TEXT_MAX_SIZE
 from test.cim.iec61970.base.core.test_identified_object import identified_object_kwargs, verify_identified_object_constructor_default, \
     verify_identified_object_constructor_kwargs, verify_identified_object_constructor_args, identified_object_args
