+++ test/cim/iec61970/base/core/test_name_type.py
@@ -8,6 +8,7 @@ from collections import Counter
 from hypothesis import given
 from hypothesis.strategies import text
 from test.cim import extract_testing_args
+from test.cim.extract_testing_args import extract_testing_args
 from test.cim.cim_creators import ALPHANUM, TEXT_MAX_SIZE
 from zepben.evolve.model.cim.iec61970.base.core.create_core_components import create_name_type
 from zepben.evolve.model.cim.iec61970.base.core.name_type import NameType
