+++ test/cim/iec61970/base/diagramlayout/test_diagram_object.py
@@ -6,7 +6,7 @@
 from hypothesis import given
 from hypothesis.strategies import lists, builds, text, floats
 
-from test.cim import extract_testing_args
+from test.cim.extract_testing_args import extract_testing_args
 from test.cim.collection_validator import validate_collection_ordered
 from test.cim.iec61970.base.core.test_identified_object import identified_object_kwargs, verify_identified_object_constructor_default, \
     verify_identified_object_constructor_kwargs, verify_identified_object_constructor_args, identified_object_args
