+++ test/cim/iec61970/base/scada/test_remote_source.py
@@ -6,7 +6,7 @@
 from hypothesis import given
 from hypothesis.strategies import builds
 
-from test.cim import extract_testing_args
+from test.cim.extract_testing_args import extract_testing_args
 from test.cim.iec61970.base.scada.test_remote_point import remote_point_kwargs, verify_remote_point_constructor_default, \
     verify_remote_point_constructor_kwargs, verify_remote_point_constructor_args, remote_point_args
 from zepben.evolve import RemoteSource, Measurement
