+++ test/cim/iec61970/base/wires/test_per_length_sequence_impedance.py
@@ -6,7 +6,7 @@
 from hypothesis import given
 from hypothesis.strategies import floats
 
-from test.cim import extract_testing_args
+from test.cim.extract_testing_args import extract_testing_args
 from test.cim.iec61970.base.wires.test_per_length_impedance import verify_per_length_impedance_constructor_default, \
     verify_per_length_impedance_constructor_kwargs, verify_per_length_impedance_constructor_args, per_length_impedance_kwargs, per_length_impedance_args
 from test.cim.cim_creators import FLOAT_MIN, FLOAT_MAX
