+++ test/cim/iec61970/base/wires/test_transformer_star_impedance.py
@@ -6,7 +6,7 @@
 from hypothesis import given
 from hypothesis.strategies import floats, builds
 
-from test.cim import extract_testing_args
+from test.cim.extract_testing_args import extract_testing_args
 from test.cim.iec61970.base.core.test_identified_object import identified_object_kwargs, verify_identified_object_constructor_default, \
     verify_identified_object_constructor_kwargs, verify_identified_object_constructor_args, identified_object_args
 from test.cim.cim_creators import FLOAT_MIN, FLOAT_MAX
