+++ test/cim/iec61970/infiec61970/feeder/test_circuit.py
@@ -7,7 +7,7 @@
 from hypothesis import given
 from hypothesis.strategies import builds, lists
 
-from test.cim import extract_testing_args
+from test.cim.extract_testing_args import extract_testing_args
 from test.cim.collection_validator import validate_collection_unordered
 from test.cim.iec61970.base.wires.test_line import verify_line_constructor_default, verify_line_constructor_kwargs, verify_line_constructor_args, line_kwargs, \
     line_args
