+++ test/cim/iec61970/infiec61970/feeder/test_loop.py
@@ -6,7 +6,7 @@
 from hypothesis import given
 from hypothesis.strategies import builds, lists
 
-from test.cim import extract_testing_args
+from test.cim.extract_testing_args import extract_testing_args
 from test.cim.collection_validator import validate_collection_unordered
 from test.cim.iec61970.base.core.test_identified_object import verify_identified_object_constructor_default, verify_identified_object_constructor_kwargs, \
     verify_identified_object_constructor_args, identified_object_kwargs, identified_object_args
